Add support for user defined growth factor.
The previous implementation was only adding a single empty
item on the buffer. It is now possible to select the
growth factor using Set_Node_List_Growth_Factor. The default
is set to zero for backward compatibility. Using a growth
factor of one can drastically improve the performances of
handling of large SOAP attachments on AWS.

Based on a patch contributed by Wiljan Derks.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/xmlada@132455 936e1b1b-40f2-da11-902a-00137254ae57