Fix handling of <choice> with elements with maxOccurs
If a <choice> has two or more elements with a maxOccurs property,
they will allow a mix of the elements, when it should only allow a
repeat of one of the elements.

The simple fix to the state machine is in schema-schema_readers.adb.
But this change resulted in a regression in the testsuite, when one
of the elements in the <choice> is a <any>. At this point, we are
trying to replace the states in the state machine, but did not
properly chain them, and therefore some states accessible via
empty transitions were not properly accessible.

The order of suggested elements in error messages was reversed in
some cases, but the error message is still value, so adjust the
baseline.

P621-020

Change-Id: I1cfb58e2fd45cfd0fc1660fa8727ab679543ba5b