use @target_alias@ for --target option and install prefix
When trying to build XmlAda for e500v2-vx6, one calls configure
with --target=e500v2-wrs-vxworks, which results in the Makefile
containing the following definition for TARGET:

    TARGET=powerpc-wrs-vxworks

This is because the --target has been canonicalized by config.sub,
and this is expected. However, this causes problems during the build,
since we then call gprbuild with --target=powerpc-wrs-vxworks
instead of the target we specified during the configure.

This patch fixes the issue by declaring a new Makefile variable
called TARGET_ALIAS, and using that for gprbuild's --target=
option, as well as the installation prefix.

Unveiled while working on O611-016.

Change-Id: I7cf3b24c48d26af0930813b1aff304dc78561aad