Fix Autoconf cross-compilation flag confusion
Previously, we used the --host and --target arguments of the configure
script to mean the build machine and the machine the package will run
on, mirroring GPRBuild's own usage of the terminology. By convention,
however, autoconf scripts usually take --build to mean the build
machine, --host to mean the machine the package will run on and
--target to mean the target of the toolchain being configured (if
relevant). Our non-standard usage tends to cause problems with some
distribution's packaging machinery (such as Void Linux's, or RHEL's),
which often follow autoconf's conventions instead.

When in Rome, do as the Romans do; hopefully this won't confuse
GPRBuild users (too much)!

Fixes #15